fix: move loading to children to menu shows
